refactor(design-catalog): DesignCatalogPage 탭 섹션 파일 분리 (#712)
1461줄 단일 파일을 catalog_tabs/ 디렉토리로 분리.
14개 섹션을 tokens/(6) + widgets/(8)로 구조화.
design_catalog_page.dart는 2줄 import + 페이지 레이아웃만 유지.

Confidence: high
Scope-risk: narrow
